switch looper from f64 to u32

+++ src/tests/shamrock/tree/key_pair_sortTests.cpp
@@ -99,7 +99,7 @@ void wrapper_bench_key_sort(std::string name) {
     logger::info_ln("ShamrockTest", "testing :", name);
 
     std::vector<f64> test_sz;
-    for (f64 i = 16; i < lim_bench; i *= 2) {
+    for (u32 i = 16; i < lim_bench; i *= 2) {
         test_sz.push_back(i);
     }
